fix(popover): refactor - useFloatingContext to platform specific file
