+++ packages/popover/src/index.tsx
@@ -1,2 +1,2 @@
-export * from './PopoverWrapper'
 export * from './Popover'
+export * from './useFloatingContext'
