+++ packages/popover/src/useFloatingContext.native.tsx
@@ -1,0 +1,4 @@
+import React, { useCallback } from 'react'
+
+// Avoid bundling react-dom into native: https://github.com/tamagui/tamagui/issues/688
+export const useFloatingContext = () => useCallback(() => {}, [])
