+++ packages/popover/src/useFloatingContext.tsx
@@ -1,0 +1,36 @@
+import type { UseFloatingProps } from '@floating-ui/react-dom-interactions'
+import {
+  useDismiss,
+  useFloating,
+  useInteractions,
+  useRole,
+} from '@floating-ui/react-dom-interactions'
+import React, { useCallback } from 'react'
+
+// Custom floating context to override the Popper on web
+export const useFloatingContext = ({ open, setOpen, breakpointActive }) =>
+  useCallback(
+    (props: UseFloatingProps) => {
+      const floating = useFloating({
+        ...props,
+        open,
+        onOpenChange: setOpen,
+      }) as any
+      const { getReferenceProps, getFloatingProps } = useInteractions([
+        // useFocus(floating.context, {
+        //   enabled: !breakpointActive,
+        //   keyboardOnly: true,
+        // }),
+        useRole(floating.context, { role: 'dialog' }),
+        useDismiss(floating.context, {
+          enabled: !breakpointActive,
+        }),
+      ])
+      return {
+        ...floating,
+        getReferenceProps,
+        getFloatingProps,
+      }
+    },
+    [open, setOpen, breakpointActive]
+  )
