+++ packages/popover/types/index.d.ts
@@ -1,3 +1,3 @@
-export * from './PopoverWrapper';
 export * from './Popover';
+export * from './useFloatingContext';
 //# sourceMappingURL=index.d.ts.map
