+++ packages/popover/types/useFloatingContext.d.ts
@@ -1,0 +1,7 @@
+import type { UseFloatingProps } from '@floating-ui/react-dom-interactions';
+export declare const useFloatingContext: ({ open, setOpen, breakpointActive }: {
+    open: any;
+    setOpen: any;
+    breakpointActive: any;
+}) => (props: UseFloatingProps) => any;
+//# sourceMappingURL=useFloatingContext.d.ts.map
