+++ packages/popover/types/useFloatingContext.native.d.ts
@@ -1,0 +1,2 @@
+export declare const useFloatingContext: () => () => void;
+//# sourceMappingURL=useFloatingContext.native.d.ts.map
